add getc and putc to primitive lisp
The most basic i/o possible. Note that they operate on the same stream
as the main interpreter loop, so the evaluation of the body containing
either command should be after the interpretor reads in a meaningful
expression. This can be achieved by placing it in part of a larger
structure that is all read in at once, for instance a cond block. That
way, whatever arm gets evaled is already loaded into memory before the
i/o starts messing with the i/o stream. You probably only want to use
this if you are going to fully commit to using just getc/putc, since
trying to mix it and interpeter stuff is going to be very painful.