partial implementation of forsp-alike
This commit contains about 60% of an implementation of forsp. It lacks
garbage collection and won't work for anything but trivial tests. This
commit is to preserve this semi-functional state before starting a
refactor of the memory system. Currently the memory is represented as
an array of 8byte values, because almost everything is meant to be
8byte aligned. But that means that the lisp-level pointers are offsets
into that array rather than byte-offset pointers of C. This is an
issue for two reasons:

A. This is weird and ugly. It doesn't actually save us any errors, and
isn't comprehensive, since we need to do weird stuff to store strings
in the dictionary and the like.

B. More damningly, this breaks the cons representation optimization,
since the idea is that for a 16-byte aligned cons cell, the bottom 4
bits of the car can be used as a tag for the interpretation of the
cell. The cons tag is zero for efficiency in scenarios where one is
confident that all the types are cells. However the proceedure tag
also has two fields that are both cons pointers. In this case one must
mask out the bottom bits to get the intended car value. However when
pointers are array offsets instead of byte offsets, then the low bits
are meaningful in the translation of 16-byte aligned values.

This could be partially fixed with less rework by making the cons and
proc tag the first two tags, so that one only needs to mask the first
bit, which is safe for 16byte values. However this rework was planned
anyway, and will be useful later when porting to bare metal.