somewhat sweeping rework for gc. Untested
Reworked a lot of the code to abide by the rules required by
new_cons. Some notational cleanup and possibly a move towards more
explicit use of the cell structure rather than semi-raw pointers is
pending. This commit exists mostly to act as a checkpoint before
starting to debug again.