Change glibc to musl libc
This is to allow us to easily and quickly build a version of libc for
this project in a scoped way. I chose musl as it fit the vision of
small and easily compilable. The reason we want to recompile libc is
to disable avx vector instructions, so that I can use gdb reverse
execution, which currently doesn't support them. Standard glibc uses
them for memcpy calls, which are included when we use printf and its
ilk. Musl also allows us to staticly compile very easily, which both
reduces the compilation time of libc and fits the vibe of this
project.